+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [4.1.4](https://github.com/riodelphino/macime/compare/v4.1.3...v4.1.4) (2026-02-20)
 
+* **docs:** Add `--log-level` option
 * **refactor:** Rename `IME_BASE` to `BASE_IME`
 * **refactor:** Refine comments, Remove unused variables
 * **refactor:** Split `macime` command fallback to func (re)